+++ Media Tagging Manager/ServiceRegistrator.cs
@@ -18,6 +18,7 @@ public sealed class ServiceRegistrator : IPluginServiceRegistrator
         serviceCollection.AddSingleton<TagBackupManager>();
         serviceCollection.AddSingleton<WatchmodeQuotaTracker>();
         serviceCollection.AddSingleton<ProviderNetworkLogoCache>();
+        serviceCollection.AddSingleton<LogoLoadStateStore>();
         serviceCollection.AddSingleton<ProviderNetworkScanner>();
         serviceCollection.AddSingleton<ManualScanRequestQueue>();
         serviceCollection.AddSingleton<MediaBrowser.Model.Tasks.IScheduledTask, RefreshAvailabilityTask>();